Use relative paths in #includes of internal dependencies
This change allows the library to be used when installed outside of the normal Arduino libraries folders. This allows the library to be bundled with a sketch so that it can be distributed as a self-contained package. Prior to this change, when the library was not installed to a standard Arduino libraries folder compilation of any sketch which included the library would fail:

C:\Users\per\AppData\Local\Temp\arduino_build_811257\sketch\src/ArduinoUnit/src/ArduinoUnitUtility/ArduinoUnitString.h:6:38: fatal error: ArduinoUnitUtility/Flash.h: No such file or directory

 #include "ArduinoUnitUtility/Flash.h"