Fix incorrect merging of PROGMEM strings
This fixes the ARDUINO_UNIT_PSTR macro to not reference the wrong
string. In particular, this makes failing macros not output "passed"
instead of "failed".

The same problem was seemingly the subject of commit 60840b8 (not all
skipped.). That commit fixed the problem for test status by not using
the ARDUINO_UNIT_PSTR macro, and seems to contain a half fix for
ARDUINO_UNIT_PSTR using __COUNTER__, but the way it was used,
__COUNTER__ was never actually expanded, so that did not help (also, it
tried to apply __COUNTER__ in the wrong place).

The actual problem *seems* to be that the `0:` label used to reference
the string constant is not actually unique. It should not need to be,
since the `0b` reference should just reference the nearest label 0,
going backwards, but I suspect that gcc ends up reordering asm blocks
and thus messes up the reference. Using a unique (but still numerical so
local) label seems to fix this. See e.g. this thread about reordering of
asm blocks:

https://gcc.gnu.org/legacy-ml/gcc-help/2017-10/msg00061.html

So, this commit reverts the __COUNTER__ bit from that commit, just using
the plain `ptr` name again (which does not need to be unique, since it
is local the the __extension__ block). It also reverts the changes
around the test status, so these now use ARDUINO_UNIT_STRING again as
originally.

Then, it applies a two-layer macro wrapper for ARDUINO_UNIT_PSTR, that
ensures that __COUNTER__ is actually expanded before being used (which
is needed when using a macro as part of a stringification or
concatenation) and ensures that __COUNTER__ is only expanded once, with
one value, rather than with different values in all three places.

Finally, this commit uses this expanded counter value to generate
unique labels, which seems to prevent this problem from occuring.